+++ frontend/src/components/Chunk.jsx
@@ -1,9 +1,10 @@
 import {useEffect,useState} from 'react'
-
+import Square from './square';
  /* 
       summary:
             n x n unit of Squares
       args:
+            id: chunk id number
             n: height and width of Chunk
             handleClick: function that handles the selection of a Square
             adjacent: state array of legal selections
@@ -11,7 +12,7 @@ import {useEffect,useState} from 'react'
             Chunk
     */
 
-export default function Chunk({n,handleClick,adjacent}) {
+export default function Chunk({id,n,handleClick,adjacent}) {
     
     const [letters, setLetters] = useState([]);
     useEffect(() => {
